Get electrofishing data for a river for one species one year

Get electrofishing data for a river for one species one year

Usage

dcf_get_efish_data(
  river,
  year = as.numeric(format(Sys.Date(), "%Y")),
  year2 = NULL,
  wanted_sites = dcf_known_efish_sites(river)
)

Arguments

river

character string with name of river.

year

numeric year to query or the string "all". Default current year

year2

optional numeric to return results between year and year2

sites

a data frame with xkoorlok and ykoorlok

Value

Return a data frame with electro fishing results. If "fiskedat" is NA the site was not fished in the time period.

Examples

if (FALSE) { # \dontrun{
res <- dcf_get_efish_data("Vindelälven", year= 2018)
head(res)
} # }