Retrieve Project Report

Retrieve the report file from a local or remote project.

Usage

dcf_report(
  project = "dissc-yale/pophive_demo",
  branch = "main",
  commit = NULL,
  provider = "https://github.com",
  cache = tempdir(),
  refresh = FALSE
)

Arguments

project

Path to a local project, or the GitHub account and repository name ("{account_name}/{repo_name}") of a remote project.

branch

Name of the remote repository branch.

commit

Commit hash; overrides branch.

provider

Base URL of the remote repository provider.

cache

Directory to store retrieved report in (at {cache}/{project}/report.json.gz).

refresh

Logical; if TRUE, will always retrieve a fresh copy of the report, even if a copy exists in cache.

Value

A data collection project report:

date	Timestamp when the report was created.

settings	The project's settings file.

source_times	A list with entries for each subproject, containing the number of milliseconds it took to run the project's scripts.

issues	A list with entries for each subproject, containing issues flagged by dcf_check, within a list with data and/or measure entries, containing character vectors of issue labels.

logs	A list with entries for each subproject, containing the logged output of their scripts.

metadata	A list with entries for each subproject, containing the datapackage of their output, as created by dcf_measure_info.

processes	A list with entries for each subproject, containing their process definitions (see dcf_add_source and/or dcf_add_bundle).

Examples

report <- dcf_report("dissc-yale/pophive_demo")
report$date
#> [1] "2026-02-27 03:02:32"
jsonlite::toJSON(report$settings, auto_unbox = TRUE, pretty = TRUE)
#> {
#>   "name": "pophive",
#>   "data_dir": "data",
#>   "github_account": "dissc-yale",
#>   "branch": "main",
#>   "repo_name": "pophive_demo"
#> }