Electofishing sites monitored in the DCF-program (WGBAST and WGNAS)

Dataset DCF::efish_sites contains information about all electrofishing sites monitored in the DCF-program. Xkoorlok and ykoorlok is the primary key to find a unique site in SERS. `DCF::WGBAST_rivers` and `DCF::WGNAS_rivers` are character vectors with rivers monitored by WG. These vectors can, for example, be used to loop do some processing for all rivers for a working group.

Usage

efish_sites

WGBAST_rivers

WGNAS_rivers

Format

WGBAST_rivers: A vector of names of Swedish WGBAST rivers. The names can be used to select data from `efish_sites`.

WGNAS_rivers: A vector of names of Swedish WGNAS rivers. The names can be used to select data from `efish_sites`.

efish_sites: A list of data frames, one for each river. Each item have is named after a river. Each data frame have the variables:

name

site name

xkoorlok

RT90 X coordinate / 10 Used as primary key in SERS

xkoorlok

RT90 Y coordinate / 10 Used as primary key in SERS

section

Name of section (blank if river is not divided in sections)

area

Estimated production area in hektar (per section if the river is divided)

used

logical Indicates if site is used in working group estimations

since

numerical First year site should be used in estimations. Zero = used since start

An object of class character of length 15.

An object of class character of length 46.

Source

https://www.slu.se/institutioner/akvatiska-resurser/databaser/elfiskeregistret/

https://www.ices.dk/community/groups/Pages/WGBAST.aspx

https://www.ices.dk/community/groups/Pages/WGNAS.aspx